secshm: fixed bitmask size

+++ secure_shmem/secshm.c
@@ -212,7 +212,7 @@ static inline int map_vm(const char *vm_name, struct vm_area_struct *vma) {
     int slot_number = page_offset / SHM_SLOT_SIZE;
     // pr_info("slot_number=0x%x page_offset=0x%lx PAGES_PER_SLOT=0x%lx\n",
     //         slot_number, page_offset, PAGES_PER_SLOT);
-    if (slot_map & (1 << slot_number))
+    if (slot_map & ((long long)1 << slot_number))
       page = pages[i]; // Normal page
     else {
       if (client_index < CLIENT_TABLE_SIZE) // Dummy page for known client
